requirements.txt
opencv-python>=4.8.0
pyautogui>=0.9.54
mss>=9.0.0
numpy>=1.24.0
Pillow>=10.0.0
keyboard>=0.13.5
easyocr>=1.7.0
pywin32>=306
